+++ cli/tests/commands/job/test_verify.py
@@ -127,7 +127,23 @@ def mock_create_sd_group(mocker):
                 {
                     "role": "roles/monitoring.metricWriter",
                     "members": ["serviceAccount:the-default-svc-account"],
-                }
+                },
+                {
+                    "role": "roles/pubsub.publisher",
+                    "members": ["serviceAccount:the-default-svc-account"],
+                },
+                {
+                    "role": "roles/pubsub.subscriber",
+                    "members": ["serviceAccount:the-default-svc-account"],
+                },
+                {
+                    "role": "roles/storage.objectCreator",
+                    "members": ["serviceAccount:the-default-svc-account"],
+                },
+                {
+                    "role": "roles/storage.objectViewer",
+                    "members": ["serviceAccount:the-default-svc-account"],
+                },
             ],
             False,
         ),
@@ -185,7 +201,7 @@ def test_verify_iam_roles(
     result = job._verify_iam_roles()
     if create_resources:
         assert (
-            "--create-resources is not able to add these roles"
+            "Klio is unable to add the required role(s)"
             in caplog.records[-1].msg
         )
 
@@ -202,6 +218,22 @@ def test_verify_iam_roles_editor(caplog, klio_config, mock_discovery_client):
     bindings = [
         {
             "role": "roles/monitoring.metricWriter",
+            "members": ["serviceAccount:the-default-svc-account"],
+        },
+        {
+            "role": "roles/pubsub.publisher",
+            "members": ["serviceAccount:the-default-svc-account"],
+        },
+        {
+            "role": "roles/pubsub.subscriber",
+            "members": ["serviceAccount:the-default-svc-account"],
+        },
+        {
+            "role": "roles/storage.objectCreator",
+            "members": ["serviceAccount:the-default-svc-account"],
+        },
+        {
+            "role": "roles/storage.objectViewer",
             "members": ["serviceAccount:the-default-svc-account"],
         },
         {
@@ -246,6 +278,22 @@ def test_verify_iam_roles_with_svc_account(klio_config, mock_discovery_client):
     bindings = [
         {
             "role": "roles/monitoring.metricWriter",
+            "members": ["serviceAccount:my.sa@something.com"],
+        },
+        {
+            "role": "roles/pubsub.publisher",
+            "members": ["serviceAccount:my.sa@something.com"],
+        },
+        {
+            "role": "roles/pubsub.subscriber",
+            "members": ["serviceAccount:my.sa@something.com"],
+        },
+        {
+            "role": "roles/storage.objectCreator",
+            "members": ["serviceAccount:my.sa@something.com"],
+        },
+        {
+            "role": "roles/storage.objectViewer",
             "members": ["serviceAccount:my.sa@something.com"],
         },
         {